Add function to select columns with specific pattern
